# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Animal Poke Backend API
  version: 1.0.0
  description: |
    Go 后端 API 契约（OpenAPI 3.1）。
    客户端仅持有 API Base URL + 设备 Token；第三方 Key 永不进入前端。
  contact:
    name: animal_poke team
  license:
    name: Proprietary
    url: https://github.com/Bluepoisons/animal_poke
servers:
  - url: https://api.animal-poke.app
    description: Production
  - url: https://api.staging.animal-poke.app
    description: Staging
...
paths:
...
  /api/v1/sync/animal:
    post:
      tags: [Sync]
      summary: Upsert single animal (idempotent)
      description: |
        Validates and persists one animal. Single and batch share the same
        validateAndSyncOne domain path. Returns stable reason_code on failure;
        never leaks raw database errors.
      operationId: syncAnimal
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncAnimalRequest"
      responses:
        "201":
          description: Synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncAnimalResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/sync/animals:
    post:
      tags: [Sync]
      summary: Batch upsert animals (non-atomic per-item results)
      description: |
        Non-atomic batch: each item is validated and synced independently via
        the same validateAndSyncOne path as POST /sync/animal. Max 100 items.
        Response is always a results array with per-item status and reason_code;
        raw DB errors are never returned.
      operationId: syncAnimalsBatch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchSyncRequest"
      responses:
        "200":
          description: Per-item batch results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchSyncResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    get:
      tags: [Sync]
      summary: Pull animals (cursor)
      description: |
        Pull animals by server_version cursor (`since_version`).
        Soft-deleted animals are returned as tombstones only
        (`uuid`, `deleted_at`, `server_version`) — never full content or precise coordinates.
      operationId: pullAnimals
      security:
        - bearerAuth: []
      parameters:
        - name: since_version
          in: query
          schema:
            type: integer
            format: int64
          description: Exclusive lower bound of server_version
        - name: cursor
          in: query
          schema:
            type: string
          description: Deprecated alias; prefer since_version
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Animals page (active rows + tombstones)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived admin JWT (AP-085). iss/aud env-isolated; includes actor/role/sid.
    adminBreakGlass:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: |
        Break-glass only. Requires X-Admin-Actor + X-Admin-Reason.
        Production rejects shared key as regular auth unless ADMIN_BREAK_GLASS_ENABLED=true.
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Deprecated alias for break-glass; prefer adminBearer.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
...
  headers:
    RequestID:
      description: Request correlation ID
      schema:
        type: string
    RetryAfter:
      description: Seconds to wait before retry
      schema:
        type: integer
  responses:
    BadRequest:
      description: Bad request
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ServiceUnavailable:
      description: Dependency unavailable
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceUnavailableError"
...
  schemas:
...
    Error:
      type: object
      required: [error, reason_code, retryable]
      description: |
        Unified API error envelope. All 4xx/5xx JSON errors should converge here.
        Handlers may attach extra diagnostic fields, but clients must rely on these core keys.
      properties:
        error:
          type: string
          description: Human-readable error message (safe for clients)
          example: invalid JSON body
        reason_code:
          type: string
          description: Machine-readable reason code
          example: bad_request
        request_id:
          type: string
          description: Correlation id (also returned as X-Request-ID header)
        retryable:
          type: boolean
          description: Whether the client may safely retry the request
          example: false
        details:
          type: object
          additionalProperties: true
          description: Optional structured context (e.g. unknown field name)
          example:
            field: extra
...
    ServiceUnavailableError:
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          required: [error, reason_code, retryable]
          properties:
            error:
              type: string
              example: service unavailable
            reason_code:
              type: string
              example: db_unavailable
            retryable:
              type: boolean
              example: true
...
    SyncAnimalRequest:
      type: object
      required: [uuid, species, rarity, generated_at]
      properties:
        uuid: { type: string, format: uuid }
        species: { type: string }
        species_label_zh: { type: string, maxLength: 64, description: Concrete Simplified Chinese species name; required for other_animal }
        breed: { type: string }
        rarity: { type: integer, minimum: 1, maximum: 5 }
        hp: { type: integer }
        atk: { type: integer }
        def: { type: integer }
        spd: { type: integer }
        class: { type: string }
        element: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        city: { type: string }
        generated_at: { type: string, format: date-time }
        inference_request_id: { type: string }
        keep_precise_location: { type: boolean }
        nickname: { type: string, maxLength: 32, description: Client-editable display nickname }
        favorite: { type: boolean, description: Favorited flag }
        locked: { type: boolean, description: Collection lock; blocks delete when true }
...
    SyncAnimalResponse:
      type: object
      required: [status, uuid]
      properties:
        status: { type: string, example: synced }
        uuid: { type: string }
        review_status: { type: string }
    BatchSyncRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/SyncAnimalRequest"
    BatchSyncResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              uuid: { type: string }
              status: { type: string }
              error: { type: string }
              reason_code: { type: string }
...